Reorg code for extractibility; add internal specs
Making it easier to extract the code and drop it into another project.

- Moved relies_on code from spec_helper to relies_on_helper
- Renamed matcher to spec/matchers/relies_on_matchers
- spec_helper now only needs to require 'relies_on_helper' and everything
  should work
- Renamed spec to spec/relies_on/internals_spec, and added more specs
- Took out lib/ altogether; moved 'my_hounds' to internals_spec
