--- lib/service.rb
@@ -1,12 +1,0 @@
-class Service
-
-  def my_hounds
-    {
-      breeding_stock: "Spartan",
-      traits: {
-        flewed_as: "Spartan",
-        sanded_as: "Spartan"
-      }
-    }
-  end
-end
